Add all Dunavant quadrature rules for the triangle
Currently, the first 7 rules of the triangle quadrature are Dunavant's
rules. The higher order rules stem from somewhere else (TODO: Add
reference when figured out)

In addition to the existing mixed rule, we added Dunavant until order
20.

The default is still the mixed rule. By wrapping the order in the type
TriangleQuadDunavant, the user can select the Dunavant rule.

TODO: There are more and better rules out there for integrating on
a triangle (some Dunavant rules have negative weights)
See, for example, the QuadPy-Package.

Probably, we want a similar package for Julia in the future.