Add all Dunavant quadrature rules for the triangle
Currently, only the first 7 rules are from Dunavant, the others were
obtained by mapping a tensor-product rule from a unit square to the
unit triangle.

TODO: There are more and better rules out there for integrating on
a triangle (some Dunavant rules have negative weights)
See, for example, the QuadPy-Package.

Probably, we want a similar package for Julia in the future.