Add support to load gmsh files via Gmsh package
This provides with load_gmsh_mesh a function that uses
the official Gmsh package to load Gmsh mesh files.

NOTE: I did not yet replace the current default gmsh reader
in the primitive functions

Advantages:
- Support more modern Gmsh mesh formats
- No need to write code for curvilinear or quadrangle/hexadral meshes

STATUS:
- The unit tests indicate full backwards compability to triangles
- So far no Gmsh parser exists regarding quadrangles/hexahedral meshes

TODO:
Currently, the gmsh3d parser for tetrahedrons performs are swap of
vertices of tetrahedrons to ensure orientation.

I am not sure this is really needed, and, given that I do not fully
grasp yet the implications regarding curvilinear meshes, did not integrate
this swap yet.

My unit tests indicate that the meshes are oriented, but I am not sure I
missed out on a more subtle point.